created seperate function files with comments to better seperate and read functions.
