+++ note.md
@@ -1,0 +1,4 @@
+echo "unset PATH
+export PATH1="/bin:/usr/bin"
+ls
+" | ./hsh
